requirements.txt
pymongo[srv]
elasticsearch==8.13.1
python-dotenv
fastapi
uvicorn[standard]
pandas